+++ src/main/java/com/maxmind/geoip2/exception/OutOfQueriesException.java
@@ -5,7 +5,6 @@ package com.maxmind.geoip2.exception;
  * remaining for the called service.
  */
 public final class OutOfQueriesException extends GeoIp2Exception {
-    private static final long serialVersionUID = 3843736987256336967L;
 
     /**
      * @param message A message explaining the cause of the error.